[mock-api] rename
